+++ meta/FFVFormFactors.m
@@ -34,6 +34,10 @@ AdjacencyGraph[FFVFormFactors`FFVGraphs[][[1]], VertexLabels -> \"Name\"]
 why such a weird numbering?
 because some procedure uses this order for fermion number flow.";
 FFVContributingDiagramsForGraph::usage = "";
+IsDiagramSupported::usage = "For the input FFV graph, determines whether 
+the input diagram is valid, by checking whether the internal structure is
+supported and the emitting fields can in fact emit the vector."
+IsChargedUnder::usage="Returns whether or not a field is charged under a given vectors gauge";
 
 Begin["Private`"];
 
@@ -51,6 +55,7 @@ FFVGraphs[] := contributingGraphs;
 EmitterL[diagram_] := diagram[[3,2]];
 EmitterR[diagram_] := diagram[[2,2]];
 Spectator[diagram_] := diagram[[2,3]];
+EmittedV[diagram_] := diagram[[3,3]];
 
 FFVContributingDiagramsForGraph[graph_, Fj_ -> {Fi_, V_}] :=
    Module[{diagrams},
@@ -65,19 +70,22 @@ FFVContributingDiagramsForGraph[graph_, Fj_ -> {Fi_, V_}] :=
    ];
 
 IsDiagramSupported[graph_, diagram_] :=
-   Module[{photonEmitter, exchangeParticle, photonEmitterAfter},
+   Module[{vectorEmitter, exchangeParticle, vectorEmitterAfter, vectorBoson},
 
-      photonEmitter = diagram[[3,2]]; (* Edge between vertices ? and ? (3rd edge of vertex 4) *)
-      photonEmitterAfter = diagram[[2,2]];
-      exchangeParticle = diagram[[2,3]]; (* Edge between vertices ? and ? (2nd edge of vertex 4) *)
+      vectorEmitter = EmitterL[diagram]; (* Edge between vertices ? and ? (3rd edge of vertex 4) *)
+      vectorEmitterAfter = EmitterR[diagram];
+      exchangeParticle = Spectator[diagram]; (* Edge between vertices ? and ? (2nd edge of vertex 4) *)
+      vectorBoson = EmittedV[diagram];
 
-      If[TreeMasses`IsFermion[photonEmitter] &&
-         TreeMasses`IsFermion[photonEmitterAfter]
-         && TreeMasses`IsScalar[exchangeParticle],
+      If[Not[IsChargedUnder[vectorEmitter,vectorBoson] && IsChargedUnder[vectorEmitterAfter,vectorBoson]],
+         Return[False];
+      ];
+
+      If[TreeMasses`IsFermion[vectorEmitter] && TreeMasses`IsFermion[vectorEmitterAfter] && TreeMasses`IsScalar[exchangeParticle],
          Return[True]
       ];
 
-      If[TreeMasses`IsFermion[exchangeParticle] && TreeMasses`IsScalar[photonEmitter] &&TreeMasses`IsScalar[photonEmitterAfter],
+      If[TreeMasses`IsFermion[exchangeParticle] && TreeMasses`IsScalar[vectorEmitter] &&TreeMasses`IsScalar[vectorEmitterAfter],
          Return[True]
       ];
 
@@ -88,6 +96,16 @@ IsDiagramSupported[graph_, diagram_] :=
          {EmitterL[diagram], EmitterR[diagram], Spectator[diagram]}, "."];
       Return[False];
    ];
+
+IsChargedUnder[field_, vector_?IsVector] := 
+  Which[(*Check 2 special cases first which are quicker*)
+    TreeMasses`IsPhoton[vector], TreeMasses`IsElectricallyCharged[field],
+    TreeMasses`IsGluon[vector],  TreeMasses`ColorChargedQ[field],
+    (*Else check that this field coupled with its anti-field can emit this vector*)
+    (*Note this will not work for vectors that couple to two different fields, e.g. W-bosons*)
+    True, SARAH`Vertex[{SARAH`AntiField[field], field, 
+        vector}, UseDependences -> True][[2, 1]] =!= 0
+  ]
 
 FFVFormFactorsCreateInterfaceFunction[Fj_ -> {Fi_, V_}, topologies_, diagrams_] :=
    Module[{prototype, definition,